Change combo and tapping terms.
I've been having issues with `sc` recently, it usually ends up not
"firing" which I'm assuming is actually triggering the "hold" (which
is command on the mac). I'm not sure whether this fixes it as I can't
build on the mac but it seems to work a little better on the xps.